Refactor: tools: Build a list of nodes in each print_* function.
Briefly, this is a list of unames for all the nodes that should have
info printed out for them.  If we aren't doing any node-based filtering,
this list will be "*".  Otherwise, it is either a provided uname or the
unames of nodes tagged with a provided tag.  It is also possible for
this list to be NULL, in which case nothing will be printed.