Refactor: libcrmcommon,libcrmservice: generalize needing stdin
services_os_action_execute() previously checked whether the agent standard
is "stonith" to know whether to pipe in standard input. This abstracts that
a bit by making it a pcmk_ra_caps value, to further move away from checking
for particular standard names and instead having a set of capabilities.