Fix: libpacemaker: Pass enough args to the resource functions.
There's a little bit of formatted output usage hiding in libpacemaker
and it uses the resource messages.  These messages now have an extra
argument that is a list of nodes whose resources can be shown, or NULL
to indicate none.  Thus, we need to pass in a list or else mysterious
crashes will occur.  Just pass in a list of all nodes.