Fix: libcrmcommon: Add another argument to pcmk__build_arg_context.
For command line programs that deal with extra args (like files or other
things on the command line that are not an option), we need a way of
describing them in the help output.  g_option_context_new has a
parameter for that as long as we pass it in.  Most of the time, this
will still be NULL.