Log: libpacemaker: check for re-promotes specifically
If a promotable clone instance is being demoted and promoted on its current
node, without also stopping and starting, it previously would be logged as
"Leave" indicating unchanged, because the current and next role are the same.

Now, check for this situation specifically, and log it as "Re-promote".

Currently, the scheduler is not capable of generating this situation, but
upcoming changes will.