Low: pacemakerd: remove unneeded IPC and CPG commands
Now that the controller handles crm_node's -l/-p needs, pacemakerd no longer
needs to. Backward compatibility isn't an issue because pacemakerd IPC isn't
proxied for Pacemaker Remote, so only local clients are relevant, and the
pacemakerd IPC never had a C API, which means it was internal only.

Without those commands, pacemakerd no longer needs to track a process mask,
connect to CPG, or maintain the peer cache.

The only remaining need for the cluster layer is to use corosync CFG to tell
corosync to initiate or block shutdown.