Fix: libstonithd: `action` parameter of RHCS-style fence-agents is shown as non-required in the metadata
The parameter `action` of RHCS-style fence-agents shouldn't usually be
specified in fencing resource configuration, although it'd be properly
remapped anyways with 9c0c3d60b4.

With this commit, the parameter `action` of RHCS-style fence-agents is
fudged to be shown as `required="0"` in the metadata output by
`stonith_admin --metadata` or `crm_resource --show-metadata`. So that
users know they don't have to configure it. Pacemaker handles and passes
the parameter to the agent when executing it.