Fix: libstonithd: `plug` parameter of RHCS-style fence-agents is shown as non-required in the metadata
With this commit, the parameter `plug` of RHCS-style fence-agents is
fudged to be shown as `required="0"` in the metadata output by
`stonith_admin --metadata` or `crm_resource --show-metadata`. So that
users know they don't have to configure it. Pacemaker handles and passes
the parameter to the agent when executing it.