Fix: tools: Quitting crm_mon shouldn't be an error.
If you ^C in interactive crm_mon, it displays the message about
terminating the connection to the daemons and then pauses for two
seconds.  That's how errors are supposed to behave.  However, just
quitting is not an error so we shouldn't be pausing there.

There's no good way to tell mon_cib_connection_destroy whether it's
being called in an error case or not because it's a callback whose
arguments are constrained elsewhere.  Instead, write two little wrapper
functions that get used as the callbacks and pass the error vs. not
parameter to the function that does the real work.