Feature: tools: Add a function to crm_mon to build a list of resources.
This is similar to what is being done for nodes, but for resources
instead.  It returns either a singleton list of the resource asked for
by name, or a list of all resources with a given tag, or a singleton
list of "*" which matches all resources.

Note: this function is marked unused for the moment, just to ensure
everything continues to build.  The attribute will be removed when the
function is used somewhere.