Build: rpm: define spec commit correctly for release candidate tags
3dd6b8c (in Pacemaker 1.1.22) broke RPM generation for release candidates.

That commit was part of a larger project to get "make dist" working, and made
"make dist" and "make export" use the same archive file name. To get this to
work, it made the rpm spec use SHORTTAG instead of TAG as the commit.

This was fine for untagged releases, but for tagged releases, distribution
naming should use the version number alone, while the spec commit should use
the entire tag.

The solution is to use separate variables for distribution naming and
the spec commit.