Fix: controller: properly detect remote node info requests
This fixes a regression introduced in 1.1.19 / 2.0.0

Ironically the original change was intended to make crm_node -n work reliably
on Pacemaker Remote nodes (even when the node name in the cluster differed from
the local hostname). However, the remote request might not be detected
appropriately depending on what the value of an uninitialized variable happened
to be.

The fix is to initialize the variable.