Fix: tools: Remove an extra space in crm_mon filtered output.
If we are filtering crm_mon output by node and give a node that doesn't
exist, an extra blank line will be output.  The node-list message knows
not to print out anything including its header, but the caller doesn't
know it won't print anything out so it adds the extra blank line between
sections.

However, the caller does know which nodes will be output.  The unames
list either has something in it (names of nodes, or "*" which means
everything) or it's an empty list because the user gave something that
didn't exist.  So we can just check to see if unames is empty.

Note that we don't do this for XML output - we want an empty list there.