Build: rpm: ensure rpmbuild knows correct tarball name
548b16b1 still wasn't complete; the update to the github URL confused rpmbuild,
which assumes the Source0 base name is the name of the tar archive. This sets
the correct base name.

Separately, ut turns out the Source0 URL has always been broken for tagged
releases. Github names official release downloads like Pacemaker-X.Y.Z.tar.gz,
and tag downloads like pacemaker-Pacemaker-X.Y.Z.tar.gz, neither of which will
work with rpmbuild, which expects pacemaker-X.Y.Z.tar.gz (which is correct for
"make rpm" builds, which are based on "make dist"/"make export"). If we changed
our release tags to start with a lowercase letter, we could configure the URL
appropriately, but that's not likely worth the effort (and inconsistency).