Fix: libpe_status: Modify filtering of inactive resources.
pe__rsc_running_on_any_node_in_list isn't exactly right for inactive
resources.  If the user passes --node= and the inactive resource could
be run on that node, we should display it.  This function only displays
inactive resources if the allowed list is "*".

Instead, add pe__rsc_allowed_on_any_node_in_list.  This function only
works for inactive resources and checks the allowed_nodes and known_on
hash tables.

These functions are separate because it's likely there's going to be a
place where we'll want to do one without the other.