Fix: libpe_status: Modify filtering of inactive resources.
Filtering inactive resources by node is a ridiculous concept.  They're
not running on any node so it doesn't really make sense to filter them
that way.

If "--inactive --node=" is given, snow all resources that are running on
the given node plus all inactive resources.  The presence of
"--inactive" means the user is requesting those resources to be shown.
Note that one consequence of this is that inactive resources will always
be shown in the XML output.