Fix: crmd: don't have all nodes clear DC's node attributes in legacy CIB mode
71e5cbd was intended to avoid the situation where the DC's shutdown attribute
stays in effect after it leaves the cluster, but paradoxically it made that
situation more likely when legacy CIB mode (used with corosync 1) is in effect.

The "fix" is to effectively revert that commit when corosync 1 is used. This
leaves the original scenario it was intended to address still a possibility,
but it's not worth trying to find a fix for corosync 1 at this stage in
pacemaker 1.1's life cycle.