Refactor: pengine,crmd: make priority-fencing-delay a compile-time option
priority-fencing-delay was backported from 2.0.4. However that could
potentially cause upgrade issues for users, since we no longer bump the feature
set version in the 1.1 series, to guarantee rolling upgrade compatility with
all 2.0 versions.

That means if a user configured priority-fencing-delay, then upgraded to a 2.0
version that didn't support it, the feature would suddenly stop working. Also,
without a feature set bump, a similar situation could occur in mixed-version
1.1 clusters.

With this commit, priority-fencing-delay will be available in a build only if
it is compiled with -DENABLE_PRIORITY_FENCING_DELAY.