Refactor: tools, libs: Add PCMK__OUTPUT_LIST_HEADER.
This macro wraps PCMK__OUTPUT_SPACER_IF, adding a way to output a list
header only if rc == pcmk_rc_no_output.  This is useful for only
outputting list headers on demand - say, in a loop over a list where
it's unknown whether any element of the list will ever actually be
printed.  In that case, this macro saves outputting a header for a list
with no items.