Refactor: scheduler: Add pe__filter_rsc_list.
This function takes a list of resources and a list of allowed resources,
returning all those actual resources present in the allowed list.  The
allowed list may be NULL, in which case an empty list will be returned,
or it can be the singleton list containing "*" in which case everything
will be returned.