Fix: tools: Add "No active resources" to one case in crm_mon output.
If no resources were displayed (possibly because we filtered on a
resource that doesn't exist), we need to add a list item to show there
were no active resources.  This is much, much easier than trying to hide
the list header in the first place.

This probably doesn't take care of every possibility, especially where
more complex resources are involved.  Those will be handled in later
patches as they come up.