Fix: tools: get_operation_list should sometimes return an empty list.
print_rsc_history takes the results of get_operation_list and performs
one further filtering on it - notify and some probe operations do not
get printed.  Thus, we can't really use the result of get_operation_list
to tell whether or not to output a header.

Instead, get_operation_list should perform that additional filtering.
If we get a non-NULL list back, then we know that we are definitely
going to print out some operation and should therefore also output a
header first.