Test: tools: Sort resources in brief output.
Due to the use of a hash table, the order of iteration may differ
between systems.  This can lead to different output in crm_mon, which in
turns may lead to test case failures.  Instead, sort the list of keys in
the resource hash table and then fetch each value always in the same
order.

This doesn't seem to happen in other places due to the use of a list -
we are adding to that list and extracting from it in the same order
every time.