Feature: libpe_status: Allow filtering bundled resources.
These are more complicated still due to the implicit resources pacemaker
creates within the bundle.  If the bundle name is given in --resource=,
everything in the bundle will be displayed.  If one of the implicit
resources is given, only it will be shown.  If a primitive resource
within the bundle is given, all instances of that resource will be
shown.