Doc: Pacemaker Explained: create sphinx template
Since converting "Pacemaker Explained" to sphinx and rST will be a huge
project, this creates a framework for tackling it chapter by chapter.

This converts the index and intro fully, but just the title of all the
chapters, prefixing the rest of the content with ".. Convert_to_RST:",
which essentially makes it a giant comment.