Feature: libpe_status, tools: Pass a list of resources to display around.
This is the first step in filtering resources by name/tag.  It is done
just like the same feature for nodes.  A list is constructed in crm_mon
and then is passed to all the resource-related output messages.

This must additionally be done in pcmk_sched_messages.c because it calls
resource output messages too.  And while I was in there, change a
g_list_append to g_list_prepend.  It shouldn't matter for a one element
list, but it's a good habit to be in.