Feature: libpe_status: Allow filtering resource groups.
These are a little more complicated than primitive resources.  If the
name of the group is given in --resource=, show the group header and
then everything underneath it, no matter the layer of nesting.  If the
name of just a member is given, show the group header and then just that
member.