Fix: cts: Use a different grep in building the cts-cli summary.
It's possible for program output to start with "* ".  crm_resource does
this in certain cases, as can anything using formatted output.  For
those cases, the summary output from cts-cli will not be what we want,
which is a "Passed: " or "Failed: " line.  So, we need to use a
different regex.