Doc: pcmk_action_timeout update
The default `pcmk_<action>_timeout` values listed in the doc are
incorrect in my experience. The monitor timeout defaults to 20s, and all
the rest default to 120s.

Also, `lib/pengine/common.c` states that `stonith-timeout` is now
unused. If so, then `pcmk_<action>_timeout` shouldn't be described as
an alternative or override to `stonith-timeout`.