Fix: libcrmcommon: Make reset and finish work more similarly.
When finish is called for HTML and XML output formats, various extra
nodes and headers are added, errors are added, etc.  None of this stuff
happens on reset.  For the HTML format, this also means things like the
CGI headers and title don't get added when reset is called.  Make these
two functions much more similar.

Regression in 2.0.3.

See: rhbz#1857728