Fix: tools, libcrmcommon: Reopen the output dest on reset.
This is needed when running crm_mon as a daemon.  When we do a reset,
we need to clear out any existing output destination and start writing
again from the beginning.  This really only matters when the destination
is a file.

The extra freopen at the end of crm_mon is to handle when crm_mon is
killed.  We need to reset the output file to its beginning before
calling finish.