Refactor: daemons, libs, tools: Use pcmk__str_null_matches where possible.
This flag allows us to replace uses of "if (s1 == NULL || crm_str_eq(s1, s2))"
with a single function call.  Note that we need to be careful about case
sensitivity only in some of these places.

A coccinelle file is provided that accomplished this refactoring and
should help to catch future uses.