Fix: tools: Change return codes of cli_resource_execute* functions.
Instead of returning one of the standard pacemaker return codes, they
should return a crm_exit_t instead.  This saves the caller having to do
the conversion, plus it allows getting rid of the is_ocf_rc checks in
crm_resource.

cli_resource_execute* were already returning an OCF error code, which is
a subset of crm_exit_t.  The other possibility was to return a
pcmk_err_generic, which was just incorrect.