Feature: libpe_status: Allow filtering primitive resources.
This is the easy case - if --resource= is given and happens to be a
primitive resource, just check to see if it's in the filter list.

I've sprinkled filtering calls throughout the rest of the resource
printing functions as well, in an effort to keep test case failures to a
minimum while I'm working on other patches.