Devel: coccinelle: Add a cocci file to catch uses of old string functions.
This file was used to perform the last several refactoring patches (plus
a little hand editing to take care of cases where the function being
refactored was in a comment).  It's added here to catch any new uses
that might pop up in the future, plus as another example of how to use
coccinelle.