Refactor: include: Use strcasecmp in crm_action_str.
This gets rid of a usage of safe_str_eq in crm/crm.h.  With that
removed, it's a lot easier to replace all uses of crm_str_eq with
pcmk__str_eq.  Otherwise, we end up having to do a lot more complicated
stuff to make sure the right includes are in the right place.