Fix: libcrmcommon: Allow default formatted output message handlers.
Sometimes, we can use a single formatted output message to handle all
available formats.  However, the registration code still requires us to
list the message repeatedly for each format.  This can be really
tedious.

Instead, add a "default" formatter.  If this is found in the
registration table, the registered function will be used for all
formats.  It's also possible to register a default, and then override
that registration on a per-formatter basis if necessary.