Feature: libcrmcommon: Add pcmk__strcmp.
This function is intended to eventually take the place of all the other
string comparison functions.  It does this through the use of flags to
control its behavior, letting it act like our other functions.

Also added is pcmk__str_eq.  This is just a quick little wrapper around
pcmk__strcmp so callers don't have to do the return value check
themselves.