High: include: Don't include an internal header in services.h.
We can't use pcmk__* in services.h because those are internal functions
and external users won't be able to include them.  Additionally, this
should be a case-sensitive comparision so use strcmp instead.

Note that the conditional is a little more verbose than it needs to be.
This is just to get around coccinelle trying to change it to use the
internal function.