Test: scheduler: update migrate-fail-9 test for migration code change
migrate-fail-7 has a failed migrate_to and successful stop on the migration
source, with only a probe on the migration target. The cluster's response is to
start the affected resource (on the source).

migrate-fail-9 is identical, except the target has a successful stop as well.
Previously, the cluster would schedule a stop and a start (on the source).
However, there's no reason a successful stop on the target should force a
redundant stop on the source. Now, it will only schedule the start.