Refactor: tools: Merge crm_mon's xml formatter into the regular one.
This code is just doing text substitutions so the XML attributes don't
have capital letters or spaces in them.  That's going to be more useful
in other tools, so merge it into the regular formatter.  This behavior
is selected by a command line argument that gets forced in crm_mon.c.