Refactor: libcrmcommon: Support quiet mode in formatted output.
First, get rid of the supports_quiet variable.  This was only used in
one place.

Second, add a new quiet variable and an is_quiet function.  Every
formatter has both of these.  The idea is that the caller sets quiet if
quiet mode is requested on the command line.  However, not every
formatter actually supports quiet mode (XML and log formatters in
particular).  So, when it's time to decide whether to print or not, the
caller uses is_quiet which also takes the formatter into account.

Third, use this in stonith_admin which is the only place supports_quiet
was being used.