Doc: all: re-style note and important boxes
The pyramid style uses a blue shade for note and warning box backgrounds,
with an exclamation point icon for note boxes and a triangle with exclamation
point for warning boxes. However it doesn't style important boxes, leaving them
looking like ordinary paragraphs (other than a barely noticeable padding).

Re-style so important boxes also have the blue shade, and use the light bulb
icon (supplied but unused by pyramid) for note, and the exclamation point for
important. (It might be better to supply our own icons, but it's not worth
the pain it would take to install them properly.)