Refactor: scheduler: Expose native_output_string as pcmk__native_output_string.
The plan is that this function can be used to help build the string that
is output for each line of "crm_resource -o" output.  It appears that
output only happens for primitive resources.  However, I've added a
check at the beginning just in case it's called for some other type of
resource.