Feature: scheduler, tools: Add a new node-and-op output message.
This is used for the output of "crm_resource -O" and "crm_resource -o".
Each operation gets printed on a single line with a brief summary of the
node it's running on preceeding it.

There's a fair bit of overlap between this and the op-history message
but there's also a number of differences.  This new message is for use
in crm_resource, and attempting to combine the two is going to change
that program's output, which may not be acceptable to some users.