Fix: scheduler: don't select instance to be promoted on a guest that can't run
The scheduler chooses clone instance(s) to be promoted from those that pass
can_be_master(). Previously, this ensured the instance was allocated to a node,
but that node might be a guest node, and its guest resource might not be
allocated to a node (i.e. can't run anywhere).

In that case, the promotion would correctly be blocked, but no other instance
would be chosen to be promoted until the next transition.

Now, ensure the guest resource is allocated as well, so another instance is
chosen.