Test: cts-scheduler: Add a test for stop-all-resources=true.
This tests the scheduler portion, but unfortunately does not test the
output/reporting portion.  cts-scheduler uses crm_simulate for output,
which currently uses different output code from crm_mon.  Being able to
test that the output of stop-all-resources=true is correct on a live
cluster will require changing crm_simulate, which is still in the works.

The XML file for this is a simplified version of cts/cli/crm_mon.xml.