Low: tools: only create validate parameters table when needed
Previously, crm_resource would always initialize the parameters hash table for
--validate, and not free it in main() since the handler for --validate would
free it. However if --validate was not called, it would be created and not
freed.

Now, it's only created when --option or --validate is specified. If --option is
specified without --validate, it will free the table. (Ideally, we'd error out
in that situation, but that will have to wait for a backward compatibility
break.)