Feature: tools: Add an output message for a list of resource names.
This is kind of an unusual output message in that it just dumps a list
of resource names with no other information.  It appears to only list
primitive resources, too.  This doesn't seem especially useful anywhere
else, so I am just adding it to crm_resource.

Note that this is one of the basic XML lists, wrapped with <list> and
<item> tags.  There's no additional useful information for the items of
this list.