Feature: tools: Use the existing resource-list message in crm_resource.
This replaces cli_resource_print_list with existing formatted output
code, getting rid of one more place where the older style output
functions are still being used.

Note that this does change the format for text output.  The older output
used indentation for displaying the members of a clone and probably
other things.  There's not really a good way to do this with the
existing text output code, short of adding more command line options for
controlling what parts of the list formatting are used.

That seems like a bit much for this one use, so instead just enable the
fancier list formatting if we are listing resources.