Refactor: tools: Use a union type for passing stonith_history around.
This lets us handle both the possibility that the history is NULL
because there is no history, as well as the possibility that it's
because there was an error in getting the history.

Note that for the moment, the union isn't fully being used.  Only the
success branch is being used.